+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Enhancements
+* authorisation - added config option to avoid adding a roles column to the users table
+  https://github.com/anvilistas/anvil-extras/pull/516
+
+
 # v2.6.1 27-Mar-2024
 
 ## Bug Fixes